Add support for recipe posts
Created a new template for recipes. This recipe is properly marked up
with all of the 'Schema.org' tags so that it will work well with search
engines and social media.

To create a new recipe, one needs only to edit the front matter.
Everything else has been set up to incorporate that front matter in the
body of the post.

easy-peasy-lemon-squeezy