Reconfigure recipe layout
Make recipe layout much more sane by off-loading all of the html markup
to the '_layout', allowing the actual body of the post to be written
with Markdown.

All the bits and pieces of the page (ingredients list, cook times, etc)
are still added via front-matter, but now the 'story' behind the recipe
can be written as an actual blog post--not a giant piece of
front-matter.